RISC-V: Add testcases disassembling "zext.h"
This commit adds testcases for "zext.h" instruction (in the 'Zbb' extension)
on the diassembler.  This instruction is a subset of either "pack" (RV32) or
"packw" (RV64) instructions (both currently in the 'Zbkb' extension) and
must be disassembled as either "zext.h", "pack" or "packw"
depending on the input.  See "zbb-zext.h-dis.s" for details per XLEN.

gas/ChangeLog:

	* testsuite/gas/riscv/zbb-zext.h-dis.s: New test for disassembling
	"zext.h" instruction.
	* testsuite/gas/riscv/zbb-zext.h-dis-32-1.d: Likewise.
	* testsuite/gas/riscv/zbb-zext.h-dis-32-2.d: Likewise.
	* testsuite/gas/riscv/zbb-zext.h-dis-32-3.d: Likewise.
	* testsuite/gas/riscv/zbb-zext.h-dis-32-4.d: Likewise.
	* testsuite/gas/riscv/zbb-zext.h-dis-64-1.d: Likewise.
	* testsuite/gas/riscv/zbb-zext.h-dis-64-2.d: Likewise.
	* testsuite/gas/riscv/zbb-zext.h-dis-64-3.d: Likewise.
	* testsuite/gas/riscv/zbb-zext.h-dis-64-4.d: Likewise.