RISC-V: Make fence.tso a non-alias
While fence.tso is a subset of original FENCE instruction in RISC-V ISA,
fence instruction as defined in the GNU Binutils (constrained to fm=0b0000)
is not a superset of fence.tso.  As a result, it resulted in ".4byte"
when disassembled with "no-aliases" option.

This commit makes fence.tso instruction a non-alias.

gas/ChangeLog:

	* testsuite/gas/riscv/fence-tso.d: Make sure that fence.tso
	is currently not an alias.

opcodes/ChangeLog:

	* riscv-opc.c (riscv_opcodes): Make fence.tso instruction
	a non-alias.