RISC-V: Split match/print steps on disassembler
For further optimization and more disassembler features, we may need to
change the core RISC-V instruction matching.  For this purpose, it is
inconvenient to have "match" and "print" steps in the same loop.

This commit rewrites riscv_disassemble_insn function so that we store
`matched_op' for matching RISC-V opcode and then print it (if that
variable is not NULL).

opcodes/ChangeLog:

	* riscv-dis.c (riscv_disassemble_insn): Split instruction
	handling to two separate steps: opcode matching and printing.