RISC-V: Use static xlen on ADDIW sequence
Because XLEN for the disassembler is computed and stored in the `xlen'
variable, this commit replaces uses of `info->mach' with `xlen'
(when testing for ADDIW / C.ADDIW address sequence).

Not just we used two ways to determine current XLEN `info->mach' and `xlen',
`xlen' is going to be more important in the future commits.

opcodes/ChangeLog:

	* riscv-dis.c (print_insn_args): Use `xlen' to determine whether
	XLEN is larger than 32.