RISC-V: Minimize arch-related initialization
`objdump' reuses the disassembler function returned by the `disassembler'
function.  That's good and benefits well from various optimizations.

However, by default, GDB (`default_print_insn' in gdb/arch-utils.c) assumes
that the `disassembler' function logic is simple and calls that function for
every instruction to be disassembled.  This is clearly a waste of time
because it probes BFD (ELF information) and re-initializes
`riscv_rps_dis' for every instruction.

To avoid repeated initialization, we cache the default architecture string
and minimize calls to the `riscv_parse_subset' function (don't call it when
the architecture string is unchanged).

For instance long `disas' command on GDB benefits from this commit.

On benchmark using GDB batch files, the author measured significant
performance improvements (27-83% on various big RISC-V programs).
Unfortunately, on interactive usecases of GDB, this improvement is rarely
observable since we don't usually disassemble such a big chunk at once and
the current disassembler is not very slow.

opcodes/ChangeLog:

	* riscv-dis.c (initial_default_arch): Special default architecture
	string which is handled separately.  (default_arch): The default
	architecture, either the ISA string from ELF attributes or
	`initial_default_arch' once initialized.  (xlen_by_arch): New.
	(update_riscv_dis_current_arch): New function to set the
	architecture string, call corresponding callback and optionally
	set `xlen_by_arch'.
	(update_riscv_dis_default_arch): New function to try to update
	the default architecture string.
	(update_riscv_dis_xlen): Consider `xlen_by_arch' when guessing
	correct XLEN.
	(init_riscv_dis_state_for_arch): New function to track whether
	the architecture string is changed.
	(init_riscv_dis_state_for_arch_and_options): Keep track of the
	architecture string change and update XLEN if it has changed.
	(riscv_get_disassembler): Update current architecture only if
	a default architecture string change is detected.  Add error
	handling.  (disassemble_init_riscv): Reset `xlen_by_arch'.