RISC-V: Use faster hash table on disassembling
This commit intends to improve performance on disassembling RISC-V code
It replaces `riscv_hash' (in opcodes/riscv-dis.c) with much faster data
structure: a sorted and partitioned hash table.

This is a technique actually used on SPARC architecture
(opcodes/sparc-dis.c) and the author simplified the algorithm even further.
Unlike SPARC, RISC-V hashed opcode table is not a table to linked lists,
it's just a table, pointing "start" elements in the sorted opcode list
(per hash code) and a global tail.

It is expected to have 14-20% performance improvements when disassembling
linked RISC-V ELF programs using `objdump'.  That is on the border of
significance and pretty nice for such a small modification
(with about 11KB heap memory allocation on 64-bit environment).

This is not the end.  This structure significantly improves plain binary
file handling (on objdump, `objdump -b binary -m riscv:rv[32|64] -D $FILE').

The author tested on various binary files including random one and big
vmlinux images and confirmed significant performance improvements (over 50%
on many cases).  This is partially due to the fact that, disassembling about
one quarter of invalid "instruction" words required iterating over one
thousand opcode entries (348 or more being vector instructions with OP-V,
that can be easily skipped with this new data structure).  Another reason
for this significance is it doesn't have various ELF overhead.

opcodes/ChangeLog:

	* riscv-dis.c (OP_HASH_LEN): Move from `riscv_disassemble_insn'.
	(OP_HASH_IDX): Move from `riscv_disassemble_insn' and mask by 0x3
	for only real 16-bit instructions.
	(compare_opcodes): New function to compare RISC-V opcode entries.
	(build_riscv_opcodes_hash_table): New function to build faster
	hash table to disassemble.
	(init_riscv_dis_state_for_arch_and_options): Build the hash table
	on the first run.
	(riscv_disassemble_insn): Use faster hash table.