RISC-V: One time CSR hash table initialization
The current disassembler intends to initialize the CSR hash table when CSR
name parsing is required at the first time. This is managed by a function-
scope static variable `init_csr' but... there's a problem.

It's never set to true.

Because of this bug, current disassembler actually initializes the CSR hash
table every time when CSR name parsing is required.  This commit sets
`init_csr' to true when the CSR hash table is parsed.

It is expected to have about 70% performance improvements when thousands of
only CSR instructions are disassembled (CSR instructions are rare in general
so real world performance improvement is not that high -- minor but
measurable, though).

opcodes/ChangeLog:

	* riscv-dis.c (print_insn_args): Make sure that CSR hash table
	initialization occurs only once.