RISC-V: Add "arch" disassembler option
This commit adds new disassembler option "arch" to override architecture
string.  Existing implementations is helpful on regular ELF files but
for instance, when debugging a binary with OpenOCD + GDB, there's a case
where such information is not available.

Also, there's a case where changing the architecture to disassemble is
helpful, even on ELF files.  For instance, Linux kernel is the prime
example.

On recent Linux kernel on RISC-V, it utilizes "alternatives" feature, which
patches the kernel on the fly, depending on the implemented extensions.  For
instance, it utilizes the 'Zicboz' extension to clear the entire page and
the 'Zbb' extension for efficient string handling.  Note that those
extension uses are determined by the runtime and at least ELF attributes
don't have such information.  Some of which (uses of 'Zicboz') are even
invisible from mapping symbols.

This commit enables objdump and GDB to use custom architecture string
instead of ELF attributes and mapping symbols (e.g. even if OpenSBI is
compiled with -march=rv64gc, we can enable H extension when disassembling
OpenSBI by specifying an architecture string: rv64gch).

It can be also helpful to test overwrapping encodings (e.g. standard hints).

gas/ChangeLog:

	* testsuite/gas/riscv/dis-arch-override.s: New arch override tests.
	* testsuite/gas/riscv/dis-arch-override-1.d: Likewise.
	* testsuite/gas/riscv/dis-arch-override-2.d: Likewise.
	* testsuite/gas/riscv/dis-arch-override-3.d: Likewise.

opcodes/ChangeLog:

	* riscv-dis.c (is_custom_arch) New.
	(update_riscv_dis_xlen): Update XLEN precedence rules.
	(set_default_riscv_dis_options): Initialize "arch".
	(parse_riscv_dis_option): Add the "arch" option.
	(riscv_get_map_state): Ignore ISA string from mapping symbols if
	custom "arch" is specified.
	(riscv_get_disassembler): Update the architecture depending on the
	custom "arch" option.
	(riscv_option_arg_t) Add RISCV_OPTION_ARG_ARCH.
	(riscv_options): Add the "arch" option.
	(disassembler_options_riscv): Add the "arch" option.