RISC-V: Break early if RISCV_GP_SYMBOL is found
If RISCV_GP_SYMBOL ("__global_pointer$") is found multiple times, it's
definately something is going wrong *for the disassembler* (note: would not
make the resulting ELF files invalid but at least confuses the disassembler)
and the disassembler will not produce the consistent result.

If so, it would be okay to just ignore second RISCV_GP_SYMBOL to make symbol
lookup slightly faster.  With this commit, the disassembler just breaks
early when a RISCV_GP_SYMBOL is found.

opcodes/ChangeLog:

	* riscv-dis.c (riscv_disassemble_insn): Break early if a
	RISCV_GP_SYMBOL is found.