RISC-V: Print using .dword on RV64
`riscv_disassemble_data' has an ability to print 8-byte data with .dword
but this is not utilized.  This is possibly because, XLEN was not set
before disassembling an instruction.

Since prerequisite commits now set XLEN in `riscv_get_disassembler', it's
now safe to use .dword on RV64.

gas/ChangeLog:

	* testsuite/gas/riscv/mapping-04b.d: Change to use RV64I.
	* testsuite/gas/riscv/mapping-norelax-04b.d: Likewise.
	* testsuite/gas/riscv/mapping-04b-32.d: New test, the same as
	mapping-04b except it uses RV32I.

opcodes/ChangeLog:

	* riscv-dis.c (riscv_data_length): Print using .dword on RV64.