RISC-V: Add mask for high address printing
This commit adds hi_addr_mask to independently determine whether a GPR
is handled as a high part of an address.  With this, we can disable handling
a GPR as a high part of an address to reduce false positives printing
addresses on the function maybe_print_address.

opcodes/ChangeLog:

	* riscv-dis.c (struct riscv_private_data): Add hi_addr_mask.
	(maybe_print_address): Use hi_addr_mask.
	(print_insn_args): Set hi_addr_mask bits.
	(init_riscv_dis_private_data): Stop using magic value of minus
	one.  Initialize hi_addr_mask.