RISC-V: New testcases for mapping symbols w/ ISA
Mapping symbols are handled per section but the "current" architecture
handled by the assembler has a file scope (regardless of the section).
We have to make sure that we can handle such text section changes while
changing the architecture.

Also, mapping symbol "$x" means that it follows the code with the default
architecture (possibly guessed by the ELF attributes).  So, if the assembler
has the default architecture, we have to make sure that "$x" is emitted.

These four new tests are meant to test such complex conditions.

gas/ChangeLog:

	* testsuite/gas/riscv/mapping-05.s:  New test for FPR switching.
	* testsuite/gas/riscv/mapping-05a.d: Likewise.
	* testsuite/gas/riscv/mapping-05b.d: Likewise.
	* testsuite/gas/riscv/mapping-06.s:  New to test FPR switching
	and make sure the last mapping symbol is the default without
	ISA string.
	* testsuite/gas/riscv/mapping-06a.d: Likewise.
	* testsuite/gas/riscv/mapping-06b.d: Likewise.
	* testsuite/gas/riscv/mapping-07.s:  New to test no excess mapping
	symbols are generated by switching between text sections.
	* testsuite/gas/riscv/mapping-07a.d: Likewise.
	* testsuite/gas/riscv/mapping-07b.d: Likewise.
	* testsuite/gas/riscv/mapping-08.s:  New to test whether the
	assembler correctly keeps track of architecture changes even if
	text sections are changed in the process.
	* testsuite/gas/riscv/mapping-08a.d: Likewise.
	* testsuite/gas/riscv/mapping-08b.d: Likewise.