RISC-V: Print top GP-relative addresses on the disassembler
This patch makes possible to print the address relative to the global
pointer even if the corresponding symbol is the highest address (0xffffffff
on RV32, 0xffffffff_ffffffff on RV64).

Despite that it is very rare to have GP the highest address, it would be
nice because we enabled highest address printing on regular cases.

opcodes/ChangeLog:

	* riscv-dis.c (struct riscv_private_data): Add `has_gp' to enable
	printing the addresses relative to GP with the highest address.
	(maybe_print_address): Utilize `has_gp'.
	(riscv_disassemble_insn): Likewise.

gas/ChangeLog:

	* testsuite/gas/riscv/dis-addr-topaddr-gp.s: New test for GP-
	relative addressing when GP is the highest address.
	* testsuite/gas/riscv/dis-addr-topaddr-gp-32.d: Likewise.
	* testsuite/gas/riscv/dis-addr-topaddr-gp-64.d: Likewise.