RISC-V: Reorganize "B"-related subextension tests
As all bit manipulation instructions are defined in multiple subextensions,
this commit will split "B" tests into separate subextension tests.

Because "B" major extension itself will not be defined, splitting and moving
these tests can be useful to avoid confusion related to non-existent
"B"-extension.

It's also better to split these because of counterintuitive grouping of
an instruction alias.  For instance, "zext.h" is a part of 'Zbb' but
"zext.w" is a specialized form of "add.uw",
which is a part of 'Zba' (RV64 only).

Separating these tests to per-subextension makes instruction tests more
complete.

gas/ChangeLog:

	* testsuite/gas/riscv/b-ext.d: Removed.
	* testsuite/gas/riscv/b-ext.s: Removed.
	* testsuite/gas/riscv/b-ext-na.d: Removed.
	* testsuite/gas/riscv/b-ext-64.d: Removed.
	* testsuite/gas/riscv/b-ext-64.s: Removed.
	* testsuite/gas/riscv/b-ext-64-na.d: Removed.
	* testsuite/gas/riscv/zba.s: Splitted from b-ext.s and b-ext-64.s.
	* testsuite/gas/riscv/zbb.s: Likewise.
	* testsuite/gas/riscv/zbc.s: Likewise.
	* testsuite/gas/riscv/zbs.s: Likewise.
	* testsuite/gas/riscv/zba.d: Splitted from b-ext.d and b-ext-64.d.
	* testsuite/gas/riscv/zbb.d: Likewise.
	* testsuite/gas/riscv/zbb-noalias.d: Likewise.
	* testsuite/gas/riscv/zbc.d: Likewise.
	* testsuite/gas/riscv/zbs.d: Likewise.
	* testsuite/gas/riscv/zbs-noalias.d: Likewise.
	* testsuite/gas/riscv/zba-64.d: Splitted from b-ext-64.d.
	* testsuite/gas/riscv/zbb-64.d: Likewise.
	* testsuite/gas/riscv/zbb-64-noalias.d: Likewise.
	* testsuite/gas/riscv/zbs-64.d: Likewise.
	* testsuite/gas/riscv/zbs-64-noalias.d: Likewise.