sim/ppc: Fix indentation on generated code
Clang generates a warning if there is a block statement and another
adjacent statement with a misleading indent ("-Wmisleading-indentation").
On the default configuration, it causes a build failure
(unless "--disable-werror" is specified).

The cause of this warning, $(builddir)/sim/ppc/semantics.c is generated by
$(srcdir)/sim/ppc/igen.c and this commit fixes generating misleading
indentation by removing two spaces.