sim/sh: Initialize some variables
Clang generates a warning if a variable may be used uninitialized on some
cases ("-Wsometimes-uninitialized").  On the default configuration, it
causes a build failure (unless "--disable-werror" is specified).

The cause of this error, $(builddir)/sim/sh/ppi.c is generated from
$(srcdir)/sim/sh/gencode.c.  Clang will detect the variable res may be used
uninitialized when used on some cases.

Likewise, GCC generates a warning if a variable may be used uninitialized
on some cases ("-Wmaybe-uninitialized").

GCC will detect variables res, res_grd, carry, overflow and greater_equal
may be used uninitialized on some cases.  Despite that GCC will not cause a
build failure even when "--enable-werror" is specified, it would be better
to fix this as well since the cause is in the same function.