sim/ppc: Remove getrusage declarations if possible
Clang generates a warning if there is a function declaration/definition
with zero arguments.  Such declarations/definitions without a prototype (an
argument list) are deprecated forms of indefinite arguments
("-Wdeprecated-non-prototype").  On the default configuration, it causes a
build failure (unless "--disable-werror" is specified).

This commit removes redundant and indefinite getrusage function declarations
if the known getrusage declaration is found.