RISC-V: Implement extension variants
If there is a instruction with multiple variants with different
requirements and the assembler fails to parse all variants, there is a case
that needs to refer ALL variants to generate proper diagnostics.

If an instruction with INSN_HAS_EXT_VARS fails on all variants, the
assembler now has a chance to modify the instruction class for proper
diagnostics.  A typical use of this feature is to select _wider_ instruction
class when necessary.

gas/ChangeLog:

	* config/tc-riscv.c (riscv_ip): Add empty INSN_HAS_EXT_VARS
	handling.

include/ChangeLog:

	* opcode/riscv.h (INSN_HAS_EXT_VARS): New.