(KNOWN_BUG) UNRATIFIED RISC-V: Add 'Zbpbo' WEXT alias macro
[DO NOT MERGE]
Until 'P' extension and its subextensions are frozen/ratified and final
version number is determined, this patch should not be merged upstream.
This commit uses unratified version 0.9 as in the documentation
(version 0.9.11).

THE AUTHOR STRONGLY RECOMMEND NOT TO MERGE THIS.

This commit adds debetable WEXT alias macro from 'Zbpbo' extension.
HOWEVER, NOT ONLY THIS COMMIT HAS A KNOWN BUG, IMPLEMENTING WEXT WITHOUT A
BUG SEEMS IMPOSSIBLE.

The spec (v0.9.11) says:

> "WEXT Rd, Rs1, Rs2" == "FSR Rd, Rs1, Rs3, Rs2"
> where Rs1 is an even register, Rs3 is an even+1 register, and Rs2[5] == 0.

But enforcing Rs2[5] == 0 is the problem.  If rd is either rs1 or rs1+1,
the result will be most likely wrong.

gas/ChangeLog:

	* config/tc-riscv.c (macro): Expand to two instructions for WEXT
	BUT THIS EXPANSION HAS A BUG.

include/ChangeLog:

	* opcode/riscv.h (M_WEXT): New for WEXT macro.

opcodes/ChangeLog:

	* riscv-opc.c (riscv_opcodes): Add WEXT macro.