sim/ppc: don't pass uninitialized value to semctl for GETVAL calls
When calling semctl with an GETVAL action, the 'union semun' argument
is unused.

Rather than passing an uninitialized variable (as we currently do),
just pass the integer 0.  This silences some compiler warnings, and is
just as correct (given the argument is ignored).

To be honest, it might be the case that the argument is not needed at
all, however, I'm a little nervous to make this change as the amount
of testing I can do of this code is limited.  Every example I found
online for a semctl/GETVAL call did pass the final argument, so that's
what I've continued to do.