sim/cris/m32c/sh: disable use of -Werror
When building the cris, m32c, and sh simulators with Clang I am seeing
build warnings from a few objects.  These three simulators currently
build with -Werror, and so these warnings cause the build to fail.

When built with gcc I don't see any warnings from these targets, so
the -Werror is fine.

As the warnings are not new, in this commit, I propose that we disable
the use of -Werror for these three simulators.  With this done it is
now possible to build the complete simulator tree using clang.