RISC-V: Add INSN_DREF to memory read/write instructions
This commit adds INSN_DREF flag (and optional size-related flag) to the
instructions that read/write the memory directly.  It however excludes
cache-related instructions that do synchronization of some sort but
otherwise don't touch the contents of the memory.

INSN_DREF and optional size flag are added to following instructions:

-   "cbo.zero" (from the 'Zicboz' extension)
-   All instructions from following custom extensions:
    -   'XTheadFMemIdx'
    -   'XTheadInt'
    -   'XTheadMemIdx'
    -   'XTheadMemPair'

opcodes/ChangeLog:

	* riscv-opc.c (riscv_opcodes): Add INSN_DREF and optional size
	flag on the instructions directly read/write memory.