RISC-V: Make T-Head testing pattern more generic
On some T-Head vendor extensions, we test against the constant
18446744073709551615 (2**64-1) to detect invalid immediate errors on -1.
However, it heavily depends on the fact that the value used to print
immediate value is a 64-bit unsigned type and this constant is not (and
should not be) important (we just want to know that -1 is not valid).

This commit replaces all such occurrences of 18446744073709551615 with
a more generic regular expression.

gas/ChangeLog:

	* testsuite/gas/riscv/x-thead-ba-fail.l: Replace
	18446744073709551615 with generic regular expression.
	* testsuite/gas/riscv/x-thead-bb-fail.l: Likewise.
	* testsuite/gas/riscv/x-thead-bs-fail.l: Likewise.
	* testsuite/gas/riscv/x-thead-fmemidx-fail.l: Likewise.
	* testsuite/gas/riscv/x-thead-memidx-fail.l: Likewise.
	* testsuite/gas/riscv/x-thead-mempair-fail.l: Likewise.