RISC-V: Support extension version number 0.0
Before this commit, the ISA string with "xextname0p0" is handled as the
'Xextname' without version number information (not version 0.0 as specified
in the string).

This commit makes able to handle version 0.0 and enable handling placeholder
extensions (with *actual* version number) easily (without being handled as
an unknown version).

This behavior is consistent with other locations that use the
"find_any_version" variable.

bfd/ChangeLog:

	* elfxx-riscv.c (riscv_parsing_subset_version): Check whether we
	have actually found a version number.