RISC-V: Renumber internal-only [GT]PREL_[IS] reloc
After ratification of the RISC-V psABI specification (version 1.0), it
is getting enhanced and improved.

Some commits include new relocation types:

[1] <https://github.com/riscv-non-isa/riscv-elf-psabi-doc/commit/eaf0b30db01255e145fd173d8a0946a4b7cce90f>
[2] <https://github.com/riscv-non-isa/riscv-elf-psabi-doc/commit/d49e48097bcad410e7e6d96ebf9d94d6d41dc9c0>

The latest draft of the RISC-V psABI specification started to use
relocation types 47-49 [2] but some of them conflict with Binutils'
internal-only relocation types:

|  N | psABI type (draft)  | Binutils type   |
| -- | ------------------- | --------------- |
| 47 | GPREL_LO12_I        | R_RISCV_GPREL_I |
| 48 | GPREL_LO12_S        | R_RISCV_GPREL_S |
| 49 | GPREL_HI20          | R_RISCV_TPREL_I |
| 50 | (reserved)          | R_RISCV_TPREL_S |

Note that R_RISCV_GPREL_[IS] cannot be used for GPREL_LO12_[IS] because
GPREL_LO12_[IS] do not allow rewrite to rd and internal R_RISCV_GPREL_[IS]
are for single instruction sequence only (that's why we have both internal
TPREL_[IS] and external TPREL_LO12_[IS]).

We have to move at least 47-49 but for locality, we move
R_RISCV_GPREL_[IS] from 47-48 to 41-42 and R_RISCV_TPREL_[IS] from 49-50
to 66-67 (note that 62-65 are reserved for other relocation types in the
latest draft of RISC-V psABI [1]).

It also notes all reserved relocation types as defined in the latest
draft of RISC-V psABI.

bfd/ChangeLog:

	* elfxx-riscv.c (howto_table): Reserve all defined relocation
	types as defined in the latest draft of RISC-V psABI.  Move
	R_RISCV_[GT]PREL_[IS] to the empty spaces.

include/ChangeLog:

	* elf/riscv.h (elf_riscv_reloc_type): Renumber
	R_RISCV_[GT]PREL_[IS] from 47-50 to 41, 42, 66 and 67.  Comment
	all defined relocation types as defined in the latest draft
	of RISC-V psABI.