RISC-V: Add implicit dependency to the XTheadVector extension
While this dependency is not directly stated in the documentation,
the XTheadVector extension cannot work without the Zicsr extension
(the documentation does not specify CSR access instruction subset
either as in the Zkr extension or the seed CSR section in the manual).

Also, making an implication to the Zicsr extension is in parity with
the ratified vector extensions (in GNU Binutils, the Zve32x extension --
a dependency of V -- depends on the Zvl32b and Zicsr extensions).

This commit adds this implicit dependency.

bfd/ChangeLog:

	* elfxx-riscv.c (riscv_implicit_subsets): Add implicit
	dependency "XTheadVector" -> "Zicsr".

gas/ChangeLog:

	* testsuite/gas/riscv/imply.s: Add implicit "XTheadVector"
	dependency to the "Zicsr" extension.
	* testsuite/gas/riscv/imply.d: Ditto.